vlib_requirements.txt
boto
grappelli
south
werkzeug
werkzeug_debugger_appengine
tastypie
dateutil
mimeparse
django_extensions
